+++ tests/test_commands/conftest.py
@@ -184,10 +184,37 @@ def base_timeseries_path():
 
 
 @pytest.fixture
+def base_timeseries_int_timestamp_path():
+    df = generate_ar_df(periods=100, start_time=10, n_segments=2, freq=None)
+    tmp = NamedTemporaryFile("w")
+    df.to_csv(tmp, index=False)
+    tmp.flush()
+    yield Path(tmp.name)
+    tmp.close()
+
+
+@pytest.fixture
 def base_timeseries_exog_path():
     df_regressors = pd.DataFrame(
         {
             "timestamp": list(pd.date_range("2021-06-01", periods=120)) * 2,
+            "regressor_1": np.arange(240),
+            "regressor_2": np.arange(240) + 5,
+            "segment": ["segment_0"] * 120 + ["segment_1"] * 120,
+        }
+    )
+    tmp = NamedTemporaryFile("w")
+    df_regressors.to_csv(tmp, index=False)
+    tmp.flush()
+    yield Path(tmp.name)
+    tmp.close()
+
+
+@pytest.fixture
+def base_timeseries_int_timestamp_exog_path():
+    df_regressors = pd.DataFrame(
+        {
+            "timestamp": np.arange(10, 130).tolist() * 2,
             "regressor_1": np.arange(240),
             "regressor_2": np.arange(240) + 5,
             "segment": ["segment_0"] * 120 + ["segment_1"] * 120,