+++ CHANGELOG.md
@@ -66,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update outliers transforms to handle integer timestamp ([#229](https://github.com/etna-team/etna/pull/229))
 - Update pipelines to handle integer timestamp ([#241](https://github.com/etna-team/etna/pull/241))
 - Add `timestamp_range` and refactor code with it ([#244](https://github.com/etna-team/etna/pull/244))
+- Update CLI to handle integer timestamp ([#246](https://github.com/etna-team/etna/pull/246))
 
 ### Fixed
 -